WIP: refactor whitelabel
