+++ app/components/whitelabel/medcryptor/medcryptor-white-label-components.js
@@ -1,7 +1,7 @@
 import {
     PAGE_NAMES,
     PAGE_COMPONENTS
-} from '../signup/signup-screens/medcryptor';
+} from '../../signup/signup-screens/medcryptor';
 
 export default {
     PAGE_NAMES,
