+++ app/components/whitelabel/peerio/peerio-white-label-components.js
@@ -1,7 +1,7 @@
 import {
     PAGE_NAMES,
     PAGE_COMPONENTS
-} from '../signup/signup-screens/peerio';
+} from '../../signup/signup-screens/peerio';
 
 export default {
     PAGE_NAMES,
