--- app/lib/whitelabel-config.js
@@ -1,5 +1,0 @@
-const whitelabel = {
-    stringReplacements: []
-};
-
-export default whitelabel;
